refactor(parameterList): moving from package parameters to lists
