+++ go/lists/base/parameter_list.go
@@ -1,10 +1,12 @@
 package base
 
 import (
+	"github.com/AssetMantle/schema/go/lists"
 	"github.com/AssetMantle/schema/go/parameters"
+	"github.com/AssetMantle/schema/go/parameters/base"
 )
 
-var _ parameters.ParameterList = (*ParameterList)(nil)
+var _ lists.ParameterList = (*ParameterList)(nil)
 
 func (parameterList *ParameterList) Get() []parameters.Parameter {
 	Parameters := make([]parameters.Parameter, len(parameterList.Parameters))
@@ -14,10 +16,10 @@ func (parameterList *ParameterList) Get() []parameters.Parameter {
 	return Parameters
 }
 
-func NewParameterList(parameters ...parameters.Parameter) parameters.ParameterList {
-	newParameters := make([]*Parameter, len(parameters))
+func NewParameterList(parameters ...parameters.Parameter) lists.ParameterList {
+	newParameters := make([]*base.Parameter, len(parameters))
 	for i, parameter := range parameters {
-		newParameters[i] = parameter.(*Parameter)
+		newParameters[i] = parameter.(*base.Parameter)
 	}
 
 	return &ParameterList{
