+++ go/lists/parameter_list.go
@@ -1,10 +1,12 @@
-package parameters
+package lists
 
 import (
 	"github.com/gogo/protobuf/proto"
+
+	"github.com/AssetMantle/schema/go/parameters"
 )
 
 type ParameterList interface {
 	proto.Message
-	Get() []Parameter
+	Get() []parameters.Parameter
 }
